requirements.txt
pandas>=2.1.0
openpyxl>=3.1.0
requests>=2.31.0
streamlit>=1.28.0
plotly>=5.18.0
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
